requirements.txt
/*************  ✨ Codeium Command 🌟  *************/
fastapi
Pillow
numpy
opencv-python-headless
/******  43c54ae4-cbb7-49e1-9767-6755c362abae  *******/